Fix for J-INTERCAL compatbility
Because comments in Google INTERCAL code are discouraged, I'll explain
this here instead. J-INTERCAL's ignorret pattern is to just restore the
variable arbitrarily, ignoring the IGNORE. Thus, this code uses a
stash/retrive that doesn't change the variable at all, but rather (in
CLC-INTERCAL) its ignorance status. C- and J-INTERCAL don't retrive
ignorance status from stashes anyway, so the whole ignorret is a no-op
for them and they proceed to change the text string from CLC-INTERCAL
to C-INTERCAL syntax.

This program might look shorter reversed, but then it doesn't work;
the issue is that C-INTERCAL errors on trying to assign out of bounds
to an ignored array. This actually fits the documentation - assigning
to ignored variables is ignored, but any side-effect of the assignment
still happens, including the bounds check - but is at least unexpected.